move cropNegativeBins to histoWithNuisances.py
